+++ salik-mobile/app/request.js
@@ -1,6 +1,6 @@
 import { View, Text } from "react-native";
 import React from "react";
-import CustomerRequests from "../components/RequestComponent/CustomerRequests";
+import CustomerRequests from "../components/CustomerRequestsComponents/CustomerRequests";
 
 const request = () => {
   return <CustomerRequests />;
